+++ src/includes/classes/Core/Utils/Html2Text.php
@@ -5,7 +5,7 @@
  * @author @jaswrks
  * @copyright WebSharks™
  */
-declare (strict_types = 1);
+declare(strict_types=1);
 namespace WebSharks\Core\Classes\Core\Utils;
 
 use WebSharks\Core\Classes;
@@ -45,9 +45,10 @@ class Html2Text extends Classes\Core\Base\Core implements Interfaces\HtmlConstan
             return $string; // Nothing to do.
         }
         $default_args = [
-            'br2nl'                 => true,
-            'strip_content_in_tags' => $this::HTML_INVISIBLE_TAGS,
-            'inject_eol_after_tags' => $this::HTML_BLOCK_TAGS,
+            'br2nl'                       => true,
+            'strip_content_in_tags'       => $this::HTML_INVISIBLE_TAGS,
+            'inject_eol_after_tags'       => $this::HTML_BLOCK_TAGS,
+            'maybe_inject_eol_after_tags' => $this::HTML_MAYBE_BLOCK_TAGS,
         ];
         $args = array_merge($default_args, $args);
         $args = array_intersect_key($args, $default_args);
@@ -55,14 +56,21 @@ class Html2Text extends Classes\Core\Base\Core implements Interfaces\HtmlConstan
         $br2nl = (bool) $args['br2nl']; // Allow line breaks?
 
         $strip_content_in_tags            = (array) $args['strip_content_in_tags'];
-        $strip_content_in_tags_regex_frag = implode('|', $this->c::escRegex($strip_content_in_tags));
+        $strip_content_in_tags_regex_frag = implode('|', $strip_content_in_tags);
 
         $inject_eol_after_tags            = (array) $args['inject_eol_after_tags'];
-        $inject_eol_after_tags_regex_frag = implode('|', $this->c::escRegex($inject_eol_after_tags));
+        $inject_eol_after_tags_regex_frag = implode('|', $inject_eol_after_tags);
+
+        $maybe_inject_eol_after_tags            = (array) $args['maybe_inject_eol_after_tags'];
+        $maybe_inject_eol_after_tags_regex_frag = implode('|', $maybe_inject_eol_after_tags);
 
         $string = preg_replace('/\<('.$strip_content_in_tags_regex_frag.')(?:\>|\s[^>]*\>).*?\<\/\\1\>/uis', '', $string);
+
         $string = preg_replace('/\<\/(?:'.$inject_eol_after_tags_regex_frag.')\>/ui', '${0}'."\n", $string);
         $string = preg_replace('/\<(?:'.$inject_eol_after_tags_regex_frag.')(?:\/\s*\>|\s[^\/>]*\/\s*\>)/ui', '${0}'."\n", $string);
+
+        $string = preg_replace('/(?:^|['."\r\n".']+)\h*\<\/(?:'.$maybe_inject_eol_after_tags_regex_frag.')\>\h*(?:['."\r\n".']+|$)/ui', '${0}'."\n", $string);
+        $string = preg_replace('/(?:^|['."\r\n".']+)\h*\<(?:'.$maybe_inject_eol_after_tags_regex_frag.')(?:\/\s*\>|\s[^\/>]*\/\s*\>)\h*(?:['."\r\n".']+|$)/ui', '${0}'."\n", $string);
 
         $string = strip_tags($string, $br2nl ? '<br>' : '');
         $string = $this->c::unescHtml($string);