+++ tests/local/markdown.php
@@ -8,4 +8,4 @@ require_once dirname(__FILE__, 2).'/includes/local.php';
 
 /* ------------------------------------------------------------------------------------------------------------------ */
 
-echo c::markdown('Testing one, two, three.', ['anchorize' => true]);
+echo c::markdown('## Testing one, two, three. {.!~@+-#foo-bar-class:1 .#foo-bar-class:100% .#1 #foo-bar-id}', ['anchorize' => true]);